fixed type issue with mol_dyn

+++ single-kernel/mol_dyn.cpp
@@ -14,7 +14,7 @@ protected:
 	int neighCount;
 	int cutsq;
 	int lj1;
-	int lj2;
+	float lj2;
 	int inum;
     BenchmarkArgs args;
